+++ pages/about.jsx
@@ -3,7 +3,7 @@
 import Meta from '@/src/components/meta/Meta'
 import Container from '@/src/components/layouts/container/Container'
 import Profile from '@/src/components/profile/Profile'
-import Hero from '@/src/components/hero/Hero'
+import Hero from '@/src/components/elements/hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/about.jpg'
