+++ pages/contact.jsx
@@ -2,7 +2,7 @@
 
 import Meta from '@/src/components/meta/Meta'
 import Container from '@/src/components/layouts/container/Container'
-import Hero from '@/src/components/hero/Hero'
+import Hero from '@/src/components/elements/hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/contact.jpg'
 import Form from '@/src/components/form/form'
