+++ pages/index.jsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { getAllPosts } from '@/libs/api'
 import Meta from '@/src/components/meta/Meta'
 import Container from '@/src/components/layouts/container/Container'
-import Hero from '@/src/components/hero/Hero'
+import Hero from '@/src/components/elements/hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/index.jpg'
 import TopAbout from '@/src/components/top-about/top-about'
