+++ src/components/elements/social/Social.jsx
@@ -1,4 +1,4 @@
-import styles from './social.module.css'
+import styles from './Social.module.css'
 import Link from 'next/link'
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
