+++ src/components/elements/table-of-contents/TableOfContents.jsx
@@ -1,6 +1,6 @@
 // 目次コンポーネント
 import { useRef, useState } from 'react'
-import styles from './table-of-contents.module.css'
+import styles from './TableOfContents.module.css'
 import { Link as Scroll } from 'react-scroll'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCircleDown } from '@fortawesome/free-solid-svg-icons'
