+++ src/components/layouts/footer/Footer.jsx
@@ -1,6 +1,6 @@
 import styles from './Footer.module.css'
-import Social from '../../social/Social'
-import Logo from '../../logo/logo'
+import Social from '../../elements/social/Social'
+import Logo from '../../elements/logo/logo'
 
 export default function Footer() {
   return (
